+++ ostap/fitting/fit2d.py
@@ -968,9 +968,9 @@ class Generic2D_pdf(PDF2) :
         ## PDF! 
         self.pdf = pdf
 
-        if not self.xvar in self.pdf.getParameters ( 0 ) : 
+        if not self.xvar in self.params () : 
             self.warning ( "Function/PDF does not depend on xvar=%s" % self.xvar.name )
-        if not self.yvar in self.pdf.getParameters ( 0 ) : 
+        if not self.yvar in self.params () : 
             self.warning ( "Function/PDF does not depend on yvar=%s" % self.yvar.name )
 
         ## add it to the list of signal components ?